demo / sync, add code from original demo unit
This is the prototype code used on the demo unit. It does mjpeg
streaming, control signal bridging to the Arduino based motor / servo
loop controller, and a couple client programs for running on a laptop to
send joystick control input and video stream playback. There's also a
diag program for reading out of of a max31865 RTD converter.

This does not include init scripts for setting up the CAN bus interfaces
or starting the video server. This only works well on a RPI3 with a
decent (external) high-gain WIFI adapter and a wide angle camera that
supports MJPEG at 320p resolution.

Code for the terminal system (laptop) can be built with cmake. This
requires sdl2, sdl2 image, jpeg, and zlib developer headers.

    cmake -DWITH_SDLPLAYER=1  ../gatorbot/

The embedded system is developed using a buildroot generated toolchain
file and has no dependencies beyond a recent compiler and standard
library.

    cmake -DCMAKE_TOOLCHAIN_FILE=/path/to/buildroot/output/host/share/buildroot/toolchainfile.cmake ../gatorbot/

It's also worth noting that default UDP settings in Linux are abysmal
when it comes to high(ish) bandwidth networking. The transport is RTP
and that uses UDP. These settings, on both systems make all the
difference.

    sysctl -w net.core.rmem_max=26214400
    sysctl -w net.core.rmem_default=26214400
    sysctl -w net.core.wmem_max=26214400
    sysctl -w net.core.wmem_default=26214400

This code is very much a work in progress, but committing now as a
synchronization point.