fix videostreamer build break
When building the SDL video player there was a build error using the
in-tree version of uvgrtp due to a header include path. Fix the player
to include the offending header using the in-tree path instead.

This was tested by building with

    cmake -DWITH_SDLPLAYER=1

Then receiving a live stream from a camera like this:

    ./recvjpegsdl /dev/video0