Add zone valve logger
The valve logger application logs state changes for up to six fluid
valves. A composite log entry representing 'system' is emitted any
time an edge is detected that indicates 'all valves off --> any valve
on' or 'any valve on --> all valves off'

This code is checked in but needs some fixups since it assumes an
add-on board implementing an 8-bit i/o expander is present. It also
makes some assumptions about how the sysfs gpio i/o lines are
enumerated. These assumptions are incorrect on more modern Linux.

Additions are needed to switch to the libgpiod API and parse a config
file to indicate which physical I/O points on the relay rack match up
with the respective GPIO pin on an add-on board or SoC.