Fixup svfload so it works on aml-s905x-cc
The GPIO pin and chip configuration are configured based on the value
that comes out of the device tree model property. This works for the
aml-s905x-cc and all raspberry pi variants provided the device tree
model contains the string 'aml-s905x-cc' and 'Raspberry' respectively.

Help output is modified to show the pin configuration for each board and
on started the detected board type, gpio pins, and gpio chip that are
used are printed to stderr.

On the amlogic SoC the JTAG toggles at around 320KHz with the 6.1
kernel. This was tested by repeatedly programming the pulsemeter design
onto a Xilinx CPLD board.